dsl: moved to the `wtf.mizu.keen.dsl` package, updated according to the API
